Cleaning code + rename eventUserStepRepository.ts
